Compute the penalty weight for the MPLE penalized likelihood method

This function computes the weight for the MPLE penalty of Moreno & Lele (2010).

computeMPLElambda(formula, data, knownOcc=numeric(0), starts,
method="BFGS",engine=c("C", "R"))

Arguments

formula
Double right-hand side formula describing covariates of detection and occupancy in that order.

data
An unmarkedFrameOccu object

knownOcc
Vector of sites that are known to be occupied. These should be supplied as row numbers of the y matrix, eg, c(3,8) if sites 3 and 8 were known to be occupied a priori.

starts
Vector of parameter starting values.

method
Optimization method used by optim.

engine
Either "C" or "R" to use fast C++ code or native R code during the optimization.

Details

See occuPEN for details and examples.

Value

The computed lambda.

References

Moreno, M. and S. R. Lele. 2010. Improved estimation of site occupancy using penalized likelihood. Ecology 91: 341-346.

Author

Rebecca A. Hutchinson

See also