2001 Delaware North American Amphibian Monitoring Program Data

frogs contains NAAMP data for Pseudacris feriarum (pfer) and Pseudacris crucifer (pcru) in 2001.

data(frogs)

Format

pcru.y

matrix of observed calling indices for pcru

pcru.bin

matrix of detections for pcru

pcru.data

array of covariates measured at the observation-level for pcru

pfer.y

matrix of observed calling indices for pfer

pfer.bin

matrix of detections for pfer

pfer.data

array of covariates measured at the observation-level for pfer

Details

The rows of pcru.y, pcru.bin, pfer.y, and pfer.bin correspond to sites and columns correspond to visits to each site. The first 2 dimensions of pfer.data and pcru.data are matrices of covariates that correspond to the observation matrices (sites \(\times\) observation), with the 3rd dimension corresponding to separate covariates.

Source

https://www.pwrc.usgs.gov/naamp/

References

Mossman MJ, Weir LA. North American Amphibian Monitoring Program (NAAMP). Amphibian Declines: the conservation status of United States species. University of California Press, Berkeley, California, USA. 2005:307-313.

Examples

data(frogs)
str(pcru.data)

#>  num [1:130, 1:3, 1:4] 37 47 57 67 77 87 97 107 117 127 ...
#>  - attr(*, "dimnames")=List of 3
#>   ..$ RouteNumStopNum: Named chr [1:130] "210302.01" "210302.02" "210302.03" "210302.04" ...
#>   .. ..- attr(*, "names")= chr [1:130] "1" "9" "21" "33" ...
#>   ..$ obsNum         : Named chr [1:3] "1" "2" "3"
#>   .. ..- attr(*, "names")= chr [1:3] "1" "5" "1367"
#>   ..$ variable       : Named chr [1:4] "MinAfterSunset" "Wind" "Sky" "Temperature"
#>   .. ..- attr(*, "names")= chr [1:4] "1" "2" "3" "4"