Distance-sampling data for the Island Scrub Jay (Aphelocoma insularis)

Data were collected at 307 survey locations ("point transects") on Santa Cruz Island, California during the Fall of 2008. The distance data are binned into 3 distance intervals [0-100], (100-200], and (200-300]. The coordinates of the survey locations as well as 3 habitat covariates are also included.

data(issj)

Format

A data frame with 307 observations on the following 8 variables.

issj[0-100]

Number of individuals detected within 100m

issj(100-200]

Detections in the interval (100-200m]

issj(200-300]

Detections in the interval (200-300m]

x

Easting (meters)

y

Northing (meters)

elevation

Elevation in meters

forest

Forest cover

chaparral

Chaparral cover

See also

References

Sillett, S. and Chandler, R.B. and Royle, J.A. and Kery, M. and Morrison, S.A. In Press. Hierarchical distance sampling models to estimate population size and habitat-specific abundance of an island endemic. Ecological Applications

Examples

data(issj)
str(issj)
#> 'data.frame':	307 obs. of  8 variables:
#>  $ issj[0-100]  : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ issj(100-200]: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ issj(200-300]: int  2 0 0 0 0 0 0 0 0 0 ...
#>  $ x            : num  234870 237083 235732 237605 234239 ...
#>  $ y            : num  3767154 3766804 3766717 3766719 3766570 ...
#>  $ elevation    : num  51.4 156.9 144.8 184.3 111.4 ...
#>  $ forest       : num  0.022054 0.006731 0.016182 0.257626 0.000716 ...
#>  $ chaparral    : num  0.242 0.466 0.769 0.206 0 ...
head(issj)
#>   issj[0-100] issj(100-200] issj(200-300]        x       y elevation
#> 1           0             0             2 234870.1 3767154   51.3851
#> 2           0             0             0 237083.0 3766804  156.8800
#> 3           0             0             0 235732.0 3766717  144.8100
#> 4           0             0             0 237605.0 3766719  184.2740
#> 5           0             0             0 234239.1 3766570  111.3530
#> 6           0             0             0 235005.1 3766420  204.1310
#>        forest chaparral
#> 1 0.022053559 0.2418731
#> 2 0.006730632 0.4664184
#> 3 0.016182157 0.7688672
#> 4 0.257625662 0.2063583
#> 5 0.000716025 0.0000000
#> 6 0.159100673 0.3402549

umf <- unmarkedFrameDS(y=as.matrix(issj[,1:3]), siteCovs=issj[,6:8],
    dist.breaks=c(0,100,200,300), unitsIn="m", survey="point")
summary(umf)
#> unmarkedFrameDS Object
#> 
#> point-transect survey design
#> Distance class cutpoints (m):  0 100 200 300 
#> 
#> 307 sites
#> Maximum number of distance classes per site: 3 
#> Mean number of distance classes per site: 3 
#> Sites with at least one detection: 76 
#> 
#> Tabulation of y observations:
#>   0   1   2   3   4   5   6   9 
#> 833  53  19   9   1   2   3   1 
#> 
#> Site-level covariates:
#>    elevation          forest           chaparral      
#>  Min.   : 16.12   Min.   :0.000000   Min.   :0.00000  
#>  1st Qu.:102.42   1st Qu.:0.000000   1st Qu.:0.06645  
#>  Median :180.94   Median :0.006731   Median :0.22326  
#>  Mean   :202.00   Mean   :0.067336   Mean   :0.27036  
#>  3rd Qu.:285.05   3rd Qu.:0.057783   3rd Qu.:0.43105  
#>  Max.   :648.22   Max.   :0.809394   Max.   :0.94401