Fit occupancy models when false positive detections occur (e.g., Royle and Link [2006] and Miller et al. [2011])

This function fits the single season occupancy model while allowing for false positive detections.

occuFP(detformula = ~ 1, FPformula = ~ 1, Bformula = ~ 1, 
stateformula = ~ 1, data, starts, method="BFGS", se = TRUE, engine = "R", ...)

Arguments

detformula
formula describing covariates of detection.

FPformula
formula describing covariates of false positive detection probability.

Bformula
formula describing covariates of probability detections are certain.

stateformula
formula describing covariates of occupancy.

data
An unmarkedFrameOccuFP object

starts
Vector of parameter starting values.

method
Optimization method used by optim.

se
Logical specifying whether or not to compute standard errors.

engine
Currently only choice is R.

...
Additional arguments to optim, such as lower and upper bounds

Details

See unmarkedFrame and unmarkedFrameOccuFP for a description of how to supply data to the data argument.

occuFP fits an extension of the standard single-season occupancy model (MacKenzie et al. 2002), which allows false positive detections. The occupancy status of a site is the same way as with the occu function, where stateformula is used to specify factors that lead to differences in occupancy probabilities among sites.

The observation process differs in that both false negative and false positive errors are modeled for observations. The function allows data to be of 3 types. These types are specified using in unmarkedFrameOccuFP as type. Occassions are specified to belong to 1 of the 3 data types and all or a subset of the data types can be combined in the same model.

For type 1 data, the detection process is assumed to fit the assumptions of the standard MacKenzie model where false negative probabilities are estimated but false positive detections are assumed not to occur. If all of your data is of this type you should use codeoccu to analyze data. The detection parameter p, which is modeled using the detformula is the only observation parameter for these data.

For type 2 data, both false negative and false positive detection probabilities are estimated. If all data is of this type the likelihood follows Royle and Link (2006). Both p (the true positive detection probability) and fp (the false positive detection probability described by fpformula) are estimated for occassions when this data type occurs

For type 3 data, observations are assumed to include both certain detections (false positives assumed not to occur) and uncertain detections that may include false positive detections. When only this data type occurs, the estimator is the same as the multiple detection state model described in Miller et al. (2011). Three observation parameters occur for this data type: p - true positive detection probability, fp - false positive detection probability, and b - the probability a true positive detection was designated as certain.

When both type 1 and type 2 data occur, the estimator is equivalent to the multiple detection method model described in Miller et al. (2011). The frog data example in the same paper uses an analysis where type 1 (dipnet surveys) and type 3 (call surveys) data were used.

Data in the y matrix of the unmarked frame should be all 0s and 1s for type 1 and type 2 data. For type 3 data, uncertain detections are given a value of 1 and certain detections a value of 2.

Value

unmarkedFitOccuFP object describing the model fit.

References

MacKenzie, D. I., J. D. Nichols, G. B. Lachman, S. Droege, J. Andrew Royle, and C. A. Langtimm. 2002. Estimating Site Occupancy Rates When Detection Probabilities Are Less Than One. Ecology 83: 2248-2255.

Miller, D.A., J.D. Nichols, B.T. McClintock, E.H.C. Grant, L.L. Bailey, and L.A. Weir. 2011. Improving occupancy estimation when two types of observational error occur: non-detection and species misidentification. Ecology 92:1422-1428.

Royle, J.A., and W.A. Link. 2006. Generalized site occupancy models allowing for false positive and false negative errors. Ecology 87:835-841.

Author

David Miller

See also

Examples

n = 100
o = 10
o1 = 5
y = matrix(0,n,o)
p = .7
r = .5
fp = 0.05
y[1:(n*.5),(o-o1+1):o] <- rbinom((n*o1*.5),1,p)
y[1:(n*.5),1:(o-o1)] <- rbinom((o-o1)*n*.5,1,r)
y[(n*.5+1):n,(o-o1+1):o] <- rbinom((n*o1*.5),1,fp)
type <- c((o-o1),o1,0)  ### vector with the number of each data type
site <- c(rep(1,n*.5*.8),rep(0,n*.5*.2),rep(1,n*.5*.2),rep(0,n*.8*.5))
occ <- matrix(c(rep(0,n*(o-o1)),rep(1,n*o1)),n,o)
site <- data.frame(habitat = site)
occ <- list(METH = occ)

umf1 <- unmarkedFrameOccuFP(y,site,occ, type = type)

m1 <- occuFP(detformula = ~ METH, FPformula = ~1,
             stateformula = ~ habitat, data = umf1)
predict(m1, type = 'fp')

#>       Predicted         SE      lower    upper
#> 1    0.07950311 0.01775005 0.05095575 0.121988
#> 2    0.07950311 0.01775005 0.05095575 0.121988
#> 3    0.07950311 0.01775005 0.05095575 0.121988
#> 4    0.07950311 0.01775005 0.05095575 0.121988
#> 5    0.07950311 0.01775005 0.05095575 0.121988
#> 6    0.07950311 0.01775005 0.05095575 0.121988
#> 7    0.07950311 0.01775005 0.05095575 0.121988
#> 8    0.07950311 0.01775005 0.05095575 0.121988
#> 9    0.07950311 0.01775005 0.05095575 0.121988
#> 10   0.07950311 0.01775005 0.05095575 0.121988
#> 11   0.07950311 0.01775005 0.05095575 0.121988
#> 12   0.07950311 0.01775005 0.05095575 0.121988
#> 13   0.07950311 0.01775005 0.05095575 0.121988
#> 14   0.07950311 0.01775005 0.05095575 0.121988
#> 15   0.07950311 0.01775005 0.05095575 0.121988
#> 16   0.07950311 0.01775005 0.05095575 0.121988
#> 17   0.07950311 0.01775005 0.05095575 0.121988
#> 18   0.07950311 0.01775005 0.05095575 0.121988
#> 19   0.07950311 0.01775005 0.05095575 0.121988
#> 20   0.07950311 0.01775005 0.05095575 0.121988
#> 21   0.07950311 0.01775005 0.05095575 0.121988
#> 22   0.07950311 0.01775005 0.05095575 0.121988
#> 23   0.07950311 0.01775005 0.05095575 0.121988
#> 24   0.07950311 0.01775005 0.05095575 0.121988
#> 25   0.07950311 0.01775005 0.05095575 0.121988
#> 26   0.07950311 0.01775005 0.05095575 0.121988
#> 27   0.07950311 0.01775005 0.05095575 0.121988
#> 28   0.07950311 0.01775005 0.05095575 0.121988
#> 29   0.07950311 0.01775005 0.05095575 0.121988
#> 30   0.07950311 0.01775005 0.05095575 0.121988
#> 31   0.07950311 0.01775005 0.05095575 0.121988
#> 32   0.07950311 0.01775005 0.05095575 0.121988
#> 33   0.07950311 0.01775005 0.05095575 0.121988
#> 34   0.07950311 0.01775005 0.05095575 0.121988
#> 35   0.07950311 0.01775005 0.05095575 0.121988
#> 36   0.07950311 0.01775005 0.05095575 0.121988
#> 37   0.07950311 0.01775005 0.05095575 0.121988
#> 38   0.07950311 0.01775005 0.05095575 0.121988
#> 39   0.07950311 0.01775005 0.05095575 0.121988
#> 40   0.07950311 0.01775005 0.05095575 0.121988
#> 41   0.07950311 0.01775005 0.05095575 0.121988
#> 42   0.07950311 0.01775005 0.05095575 0.121988
#> 43   0.07950311 0.01775005 0.05095575 0.121988
#> 44   0.07950311 0.01775005 0.05095575 0.121988
#> 45   0.07950311 0.01775005 0.05095575 0.121988
#> 46   0.07950311 0.01775005 0.05095575 0.121988
#> 47   0.07950311 0.01775005 0.05095575 0.121988
#> 48   0.07950311 0.01775005 0.05095575 0.121988
#> 49   0.07950311 0.01775005 0.05095575 0.121988
#> 50   0.07950311 0.01775005 0.05095575 0.121988
#> 51   0.07950311 0.01775005 0.05095575 0.121988
#> 52   0.07950311 0.01775005 0.05095575 0.121988
#> 53   0.07950311 0.01775005 0.05095575 0.121988
#> 54   0.07950311 0.01775005 0.05095575 0.121988
#> 55   0.07950311 0.01775005 0.05095575 0.121988
#> 56   0.07950311 0.01775005 0.05095575 0.121988
#> 57   0.07950311 0.01775005 0.05095575 0.121988
#> 58   0.07950311 0.01775005 0.05095575 0.121988
#> 59   0.07950311 0.01775005 0.05095575 0.121988
#> 60   0.07950311 0.01775005 0.05095575 0.121988
#> 61   0.07950311 0.01775005 0.05095575 0.121988
#> 62   0.07950311 0.01775005 0.05095575 0.121988
#> 63   0.07950311 0.01775005 0.05095575 0.121988
#> 64   0.07950311 0.01775005 0.05095575 0.121988
#> 65   0.07950311 0.01775005 0.05095575 0.121988
#> 66   0.07950311 0.01775005 0.05095575 0.121988
#> 67   0.07950311 0.01775005 0.05095575 0.121988
#> 68   0.07950311 0.01775005 0.05095575 0.121988
#> 69   0.07950311 0.01775005 0.05095575 0.121988
#> 70   0.07950311 0.01775005 0.05095575 0.121988
#> 71   0.07950311 0.01775005 0.05095575 0.121988
#> 72   0.07950311 0.01775005 0.05095575 0.121988
#> 73   0.07950311 0.01775005 0.05095575 0.121988
#> 74   0.07950311 0.01775005 0.05095575 0.121988
#> 75   0.07950311 0.01775005 0.05095575 0.121988
#> 76   0.07950311 0.01775005 0.05095575 0.121988
#> 77   0.07950311 0.01775005 0.05095575 0.121988
#> 78   0.07950311 0.01775005 0.05095575 0.121988
#> 79   0.07950311 0.01775005 0.05095575 0.121988
#> 80   0.07950311 0.01775005 0.05095575 0.121988
#> 81   0.07950311 0.01775005 0.05095575 0.121988
#> 82   0.07950311 0.01775005 0.05095575 0.121988
#> 83   0.07950311 0.01775005 0.05095575 0.121988
#> 84   0.07950311 0.01775005 0.05095575 0.121988
#> 85   0.07950311 0.01775005 0.05095575 0.121988
#> 86   0.07950311 0.01775005 0.05095575 0.121988
#> 87   0.07950311 0.01775005 0.05095575 0.121988
#> 88   0.07950311 0.01775005 0.05095575 0.121988
#> 89   0.07950311 0.01775005 0.05095575 0.121988
#> 90   0.07950311 0.01775005 0.05095575 0.121988
#> 91   0.07950311 0.01775005 0.05095575 0.121988
#> 92   0.07950311 0.01775005 0.05095575 0.121988
#> 93   0.07950311 0.01775005 0.05095575 0.121988
#> 94   0.07950311 0.01775005 0.05095575 0.121988
#> 95   0.07950311 0.01775005 0.05095575 0.121988
#> 96   0.07950311 0.01775005 0.05095575 0.121988
#> 97   0.07950311 0.01775005 0.05095575 0.121988
#> 98   0.07950311 0.01775005 0.05095575 0.121988
#> 99   0.07950311 0.01775005 0.05095575 0.121988
#> 100  0.07950311 0.01775005 0.05095575 0.121988
#> 101  0.07950311 0.01775005 0.05095575 0.121988
#> 102  0.07950311 0.01775005 0.05095575 0.121988
#> 103  0.07950311 0.01775005 0.05095575 0.121988
#> 104  0.07950311 0.01775005 0.05095575 0.121988
#> 105  0.07950311 0.01775005 0.05095575 0.121988
#> 106  0.07950311 0.01775005 0.05095575 0.121988
#> 107  0.07950311 0.01775005 0.05095575 0.121988
#> 108  0.07950311 0.01775005 0.05095575 0.121988
#> 109  0.07950311 0.01775005 0.05095575 0.121988
#> 110  0.07950311 0.01775005 0.05095575 0.121988
#> 111  0.07950311 0.01775005 0.05095575 0.121988
#> 112  0.07950311 0.01775005 0.05095575 0.121988
#> 113  0.07950311 0.01775005 0.05095575 0.121988
#> 114  0.07950311 0.01775005 0.05095575 0.121988
#> 115  0.07950311 0.01775005 0.05095575 0.121988
#> 116  0.07950311 0.01775005 0.05095575 0.121988
#> 117  0.07950311 0.01775005 0.05095575 0.121988
#> 118  0.07950311 0.01775005 0.05095575 0.121988
#> 119  0.07950311 0.01775005 0.05095575 0.121988
#> 120  0.07950311 0.01775005 0.05095575 0.121988
#> 121  0.07950311 0.01775005 0.05095575 0.121988
#> 122  0.07950311 0.01775005 0.05095575 0.121988
#> 123  0.07950311 0.01775005 0.05095575 0.121988
#> 124  0.07950311 0.01775005 0.05095575 0.121988
#> 125  0.07950311 0.01775005 0.05095575 0.121988
#> 126  0.07950311 0.01775005 0.05095575 0.121988
#> 127  0.07950311 0.01775005 0.05095575 0.121988
#> 128  0.07950311 0.01775005 0.05095575 0.121988
#> 129  0.07950311 0.01775005 0.05095575 0.121988
#> 130  0.07950311 0.01775005 0.05095575 0.121988
#> 131  0.07950311 0.01775005 0.05095575 0.121988
#> 132  0.07950311 0.01775005 0.05095575 0.121988
#> 133  0.07950311 0.01775005 0.05095575 0.121988
#> 134  0.07950311 0.01775005 0.05095575 0.121988
#> 135  0.07950311 0.01775005 0.05095575 0.121988
#> 136  0.07950311 0.01775005 0.05095575 0.121988
#> 137  0.07950311 0.01775005 0.05095575 0.121988
#> 138  0.07950311 0.01775005 0.05095575 0.121988
#> 139  0.07950311 0.01775005 0.05095575 0.121988
#> 140  0.07950311 0.01775005 0.05095575 0.121988
#> 141  0.07950311 0.01775005 0.05095575 0.121988
#> 142  0.07950311 0.01775005 0.05095575 0.121988
#> 143  0.07950311 0.01775005 0.05095575 0.121988
#> 144  0.07950311 0.01775005 0.05095575 0.121988
#> 145  0.07950311 0.01775005 0.05095575 0.121988
#> 146  0.07950311 0.01775005 0.05095575 0.121988
#> 147  0.07950311 0.01775005 0.05095575 0.121988
#> 148  0.07950311 0.01775005 0.05095575 0.121988
#> 149  0.07950311 0.01775005 0.05095575 0.121988
#> 150  0.07950311 0.01775005 0.05095575 0.121988
#> 151  0.07950311 0.01775005 0.05095575 0.121988
#> 152  0.07950311 0.01775005 0.05095575 0.121988
#> 153  0.07950311 0.01775005 0.05095575 0.121988
#> 154  0.07950311 0.01775005 0.05095575 0.121988
#> 155  0.07950311 0.01775005 0.05095575 0.121988
#> 156  0.07950311 0.01775005 0.05095575 0.121988
#> 157  0.07950311 0.01775005 0.05095575 0.121988
#> 158  0.07950311 0.01775005 0.05095575 0.121988
#> 159  0.07950311 0.01775005 0.05095575 0.121988
#> 160  0.07950311 0.01775005 0.05095575 0.121988
#> 161  0.07950311 0.01775005 0.05095575 0.121988
#> 162  0.07950311 0.01775005 0.05095575 0.121988
#> 163  0.07950311 0.01775005 0.05095575 0.121988
#> 164  0.07950311 0.01775005 0.05095575 0.121988
#> 165  0.07950311 0.01775005 0.05095575 0.121988
#> 166  0.07950311 0.01775005 0.05095575 0.121988
#> 167  0.07950311 0.01775005 0.05095575 0.121988
#> 168  0.07950311 0.01775005 0.05095575 0.121988
#> 169  0.07950311 0.01775005 0.05095575 0.121988
#> 170  0.07950311 0.01775005 0.05095575 0.121988
#> 171  0.07950311 0.01775005 0.05095575 0.121988
#> 172  0.07950311 0.01775005 0.05095575 0.121988
#> 173  0.07950311 0.01775005 0.05095575 0.121988
#> 174  0.07950311 0.01775005 0.05095575 0.121988
#> 175  0.07950311 0.01775005 0.05095575 0.121988
#> 176  0.07950311 0.01775005 0.05095575 0.121988
#> 177  0.07950311 0.01775005 0.05095575 0.121988
#> 178  0.07950311 0.01775005 0.05095575 0.121988
#> 179  0.07950311 0.01775005 0.05095575 0.121988
#> 180  0.07950311 0.01775005 0.05095575 0.121988
#> 181  0.07950311 0.01775005 0.05095575 0.121988
#> 182  0.07950311 0.01775005 0.05095575 0.121988
#> 183  0.07950311 0.01775005 0.05095575 0.121988
#> 184  0.07950311 0.01775005 0.05095575 0.121988
#> 185  0.07950311 0.01775005 0.05095575 0.121988
#> 186  0.07950311 0.01775005 0.05095575 0.121988
#> 187  0.07950311 0.01775005 0.05095575 0.121988
#> 188  0.07950311 0.01775005 0.05095575 0.121988
#> 189  0.07950311 0.01775005 0.05095575 0.121988
#> 190  0.07950311 0.01775005 0.05095575 0.121988
#> 191  0.07950311 0.01775005 0.05095575 0.121988
#> 192  0.07950311 0.01775005 0.05095575 0.121988
#> 193  0.07950311 0.01775005 0.05095575 0.121988
#> 194  0.07950311 0.01775005 0.05095575 0.121988
#> 195  0.07950311 0.01775005 0.05095575 0.121988
#> 196  0.07950311 0.01775005 0.05095575 0.121988
#> 197  0.07950311 0.01775005 0.05095575 0.121988
#> 198  0.07950311 0.01775005 0.05095575 0.121988
#> 199  0.07950311 0.01775005 0.05095575 0.121988
#> 200  0.07950311 0.01775005 0.05095575 0.121988
#> 201  0.07950311 0.01775005 0.05095575 0.121988
#> 202  0.07950311 0.01775005 0.05095575 0.121988
#> 203  0.07950311 0.01775005 0.05095575 0.121988
#> 204  0.07950311 0.01775005 0.05095575 0.121988
#> 205  0.07950311 0.01775005 0.05095575 0.121988
#> 206  0.07950311 0.01775005 0.05095575 0.121988
#> 207  0.07950311 0.01775005 0.05095575 0.121988
#> 208  0.07950311 0.01775005 0.05095575 0.121988
#> 209  0.07950311 0.01775005 0.05095575 0.121988
#> 210  0.07950311 0.01775005 0.05095575 0.121988
#> 211  0.07950311 0.01775005 0.05095575 0.121988
#> 212  0.07950311 0.01775005 0.05095575 0.121988
#> 213  0.07950311 0.01775005 0.05095575 0.121988
#> 214  0.07950311 0.01775005 0.05095575 0.121988
#> 215  0.07950311 0.01775005 0.05095575 0.121988
#> 216  0.07950311 0.01775005 0.05095575 0.121988
#> 217  0.07950311 0.01775005 0.05095575 0.121988
#> 218  0.07950311 0.01775005 0.05095575 0.121988
#> 219  0.07950311 0.01775005 0.05095575 0.121988
#> 220  0.07950311 0.01775005 0.05095575 0.121988
#> 221  0.07950311 0.01775005 0.05095575 0.121988
#> 222  0.07950311 0.01775005 0.05095575 0.121988
#> 223  0.07950311 0.01775005 0.05095575 0.121988
#> 224  0.07950311 0.01775005 0.05095575 0.121988
#> 225  0.07950311 0.01775005 0.05095575 0.121988
#> 226  0.07950311 0.01775005 0.05095575 0.121988
#> 227  0.07950311 0.01775005 0.05095575 0.121988
#> 228  0.07950311 0.01775005 0.05095575 0.121988
#> 229  0.07950311 0.01775005 0.05095575 0.121988
#> 230  0.07950311 0.01775005 0.05095575 0.121988
#> 231  0.07950311 0.01775005 0.05095575 0.121988
#> 232  0.07950311 0.01775005 0.05095575 0.121988
#> 233  0.07950311 0.01775005 0.05095575 0.121988
#> 234  0.07950311 0.01775005 0.05095575 0.121988
#> 235  0.07950311 0.01775005 0.05095575 0.121988
#> 236  0.07950311 0.01775005 0.05095575 0.121988
#> 237  0.07950311 0.01775005 0.05095575 0.121988
#> 238  0.07950311 0.01775005 0.05095575 0.121988
#> 239  0.07950311 0.01775005 0.05095575 0.121988
#> 240  0.07950311 0.01775005 0.05095575 0.121988
#> 241  0.07950311 0.01775005 0.05095575 0.121988
#> 242  0.07950311 0.01775005 0.05095575 0.121988
#> 243  0.07950311 0.01775005 0.05095575 0.121988
#> 244  0.07950311 0.01775005 0.05095575 0.121988
#> 245  0.07950311 0.01775005 0.05095575 0.121988
#> 246  0.07950311 0.01775005 0.05095575 0.121988
#> 247  0.07950311 0.01775005 0.05095575 0.121988
#> 248  0.07950311 0.01775005 0.05095575 0.121988
#> 249  0.07950311 0.01775005 0.05095575 0.121988
#> 250  0.07950311 0.01775005 0.05095575 0.121988
#> 251  0.07950311 0.01775005 0.05095575 0.121988
#> 252  0.07950311 0.01775005 0.05095575 0.121988
#> 253  0.07950311 0.01775005 0.05095575 0.121988
#> 254  0.07950311 0.01775005 0.05095575 0.121988
#> 255  0.07950311 0.01775005 0.05095575 0.121988
#> 256  0.07950311 0.01775005 0.05095575 0.121988
#> 257  0.07950311 0.01775005 0.05095575 0.121988
#> 258  0.07950311 0.01775005 0.05095575 0.121988
#> 259  0.07950311 0.01775005 0.05095575 0.121988
#> 260  0.07950311 0.01775005 0.05095575 0.121988
#> 261  0.07950311 0.01775005 0.05095575 0.121988
#> 262  0.07950311 0.01775005 0.05095575 0.121988
#> 263  0.07950311 0.01775005 0.05095575 0.121988
#> 264  0.07950311 0.01775005 0.05095575 0.121988
#> 265  0.07950311 0.01775005 0.05095575 0.121988
#> 266  0.07950311 0.01775005 0.05095575 0.121988
#> 267  0.07950311 0.01775005 0.05095575 0.121988
#> 268  0.07950311 0.01775005 0.05095575 0.121988
#> 269  0.07950311 0.01775005 0.05095575 0.121988
#> 270  0.07950311 0.01775005 0.05095575 0.121988
#> 271  0.07950311 0.01775005 0.05095575 0.121988
#> 272  0.07950311 0.01775005 0.05095575 0.121988
#> 273  0.07950311 0.01775005 0.05095575 0.121988
#> 274  0.07950311 0.01775005 0.05095575 0.121988
#> 275  0.07950311 0.01775005 0.05095575 0.121988
#> 276  0.07950311 0.01775005 0.05095575 0.121988
#> 277  0.07950311 0.01775005 0.05095575 0.121988
#> 278  0.07950311 0.01775005 0.05095575 0.121988
#> 279  0.07950311 0.01775005 0.05095575 0.121988
#> 280  0.07950311 0.01775005 0.05095575 0.121988
#> 281  0.07950311 0.01775005 0.05095575 0.121988
#> 282  0.07950311 0.01775005 0.05095575 0.121988
#> 283  0.07950311 0.01775005 0.05095575 0.121988
#> 284  0.07950311 0.01775005 0.05095575 0.121988
#> 285  0.07950311 0.01775005 0.05095575 0.121988
#> 286  0.07950311 0.01775005 0.05095575 0.121988
#> 287  0.07950311 0.01775005 0.05095575 0.121988
#> 288  0.07950311 0.01775005 0.05095575 0.121988
#> 289  0.07950311 0.01775005 0.05095575 0.121988
#> 290  0.07950311 0.01775005 0.05095575 0.121988
#> 291  0.07950311 0.01775005 0.05095575 0.121988
#> 292  0.07950311 0.01775005 0.05095575 0.121988
#> 293  0.07950311 0.01775005 0.05095575 0.121988
#> 294  0.07950311 0.01775005 0.05095575 0.121988
#> 295  0.07950311 0.01775005 0.05095575 0.121988
#> 296  0.07950311 0.01775005 0.05095575 0.121988
#> 297  0.07950311 0.01775005 0.05095575 0.121988
#> 298  0.07950311 0.01775005 0.05095575 0.121988
#> 299  0.07950311 0.01775005 0.05095575 0.121988
#> 300  0.07950311 0.01775005 0.05095575 0.121988
#> 301  0.07950311 0.01775005 0.05095575 0.121988
#> 302  0.07950311 0.01775005 0.05095575 0.121988
#> 303  0.07950311 0.01775005 0.05095575 0.121988
#> 304  0.07950311 0.01775005 0.05095575 0.121988
#> 305  0.07950311 0.01775005 0.05095575 0.121988
#> 306  0.07950311 0.01775005 0.05095575 0.121988
#> 307  0.07950311 0.01775005 0.05095575 0.121988
#> 308  0.07950311 0.01775005 0.05095575 0.121988
#> 309  0.07950311 0.01775005 0.05095575 0.121988
#> 310  0.07950311 0.01775005 0.05095575 0.121988
#> 311  0.07950311 0.01775005 0.05095575 0.121988
#> 312  0.07950311 0.01775005 0.05095575 0.121988
#> 313  0.07950311 0.01775005 0.05095575 0.121988
#> 314  0.07950311 0.01775005 0.05095575 0.121988
#> 315  0.07950311 0.01775005 0.05095575 0.121988
#> 316  0.07950311 0.01775005 0.05095575 0.121988
#> 317  0.07950311 0.01775005 0.05095575 0.121988
#> 318  0.07950311 0.01775005 0.05095575 0.121988
#> 319  0.07950311 0.01775005 0.05095575 0.121988
#> 320  0.07950311 0.01775005 0.05095575 0.121988
#> 321  0.07950311 0.01775005 0.05095575 0.121988
#> 322  0.07950311 0.01775005 0.05095575 0.121988
#> 323  0.07950311 0.01775005 0.05095575 0.121988
#> 324  0.07950311 0.01775005 0.05095575 0.121988
#> 325  0.07950311 0.01775005 0.05095575 0.121988
#> 326  0.07950311 0.01775005 0.05095575 0.121988
#> 327  0.07950311 0.01775005 0.05095575 0.121988
#> 328  0.07950311 0.01775005 0.05095575 0.121988
#> 329  0.07950311 0.01775005 0.05095575 0.121988
#> 330  0.07950311 0.01775005 0.05095575 0.121988
#> 331  0.07950311 0.01775005 0.05095575 0.121988
#> 332  0.07950311 0.01775005 0.05095575 0.121988
#> 333  0.07950311 0.01775005 0.05095575 0.121988
#> 334  0.07950311 0.01775005 0.05095575 0.121988
#> 335  0.07950311 0.01775005 0.05095575 0.121988
#> 336  0.07950311 0.01775005 0.05095575 0.121988
#> 337  0.07950311 0.01775005 0.05095575 0.121988
#> 338  0.07950311 0.01775005 0.05095575 0.121988
#> 339  0.07950311 0.01775005 0.05095575 0.121988
#> 340  0.07950311 0.01775005 0.05095575 0.121988
#> 341  0.07950311 0.01775005 0.05095575 0.121988
#> 342  0.07950311 0.01775005 0.05095575 0.121988
#> 343  0.07950311 0.01775005 0.05095575 0.121988
#> 344  0.07950311 0.01775005 0.05095575 0.121988
#> 345  0.07950311 0.01775005 0.05095575 0.121988
#> 346  0.07950311 0.01775005 0.05095575 0.121988
#> 347  0.07950311 0.01775005 0.05095575 0.121988
#> 348  0.07950311 0.01775005 0.05095575 0.121988
#> 349  0.07950311 0.01775005 0.05095575 0.121988
#> 350  0.07950311 0.01775005 0.05095575 0.121988
#> 351  0.07950311 0.01775005 0.05095575 0.121988
#> 352  0.07950311 0.01775005 0.05095575 0.121988
#> 353  0.07950311 0.01775005 0.05095575 0.121988
#> 354  0.07950311 0.01775005 0.05095575 0.121988
#> 355  0.07950311 0.01775005 0.05095575 0.121988
#> 356  0.07950311 0.01775005 0.05095575 0.121988
#> 357  0.07950311 0.01775005 0.05095575 0.121988
#> 358  0.07950311 0.01775005 0.05095575 0.121988
#> 359  0.07950311 0.01775005 0.05095575 0.121988
#> 360  0.07950311 0.01775005 0.05095575 0.121988
#> 361  0.07950311 0.01775005 0.05095575 0.121988
#> 362  0.07950311 0.01775005 0.05095575 0.121988
#> 363  0.07950311 0.01775005 0.05095575 0.121988
#> 364  0.07950311 0.01775005 0.05095575 0.121988
#> 365  0.07950311 0.01775005 0.05095575 0.121988
#> 366  0.07950311 0.01775005 0.05095575 0.121988
#> 367  0.07950311 0.01775005 0.05095575 0.121988
#> 368  0.07950311 0.01775005 0.05095575 0.121988
#> 369  0.07950311 0.01775005 0.05095575 0.121988
#> 370  0.07950311 0.01775005 0.05095575 0.121988
#> 371  0.07950311 0.01775005 0.05095575 0.121988
#> 372  0.07950311 0.01775005 0.05095575 0.121988
#> 373  0.07950311 0.01775005 0.05095575 0.121988
#> 374  0.07950311 0.01775005 0.05095575 0.121988
#> 375  0.07950311 0.01775005 0.05095575 0.121988
#> 376  0.07950311 0.01775005 0.05095575 0.121988
#> 377  0.07950311 0.01775005 0.05095575 0.121988
#> 378  0.07950311 0.01775005 0.05095575 0.121988
#> 379  0.07950311 0.01775005 0.05095575 0.121988
#> 380  0.07950311 0.01775005 0.05095575 0.121988
#> 381  0.07950311 0.01775005 0.05095575 0.121988
#> 382  0.07950311 0.01775005 0.05095575 0.121988
#> 383  0.07950311 0.01775005 0.05095575 0.121988
#> 384  0.07950311 0.01775005 0.05095575 0.121988
#> 385  0.07950311 0.01775005 0.05095575 0.121988
#> 386  0.07950311 0.01775005 0.05095575 0.121988
#> 387  0.07950311 0.01775005 0.05095575 0.121988
#> 388  0.07950311 0.01775005 0.05095575 0.121988
#> 389  0.07950311 0.01775005 0.05095575 0.121988
#> 390  0.07950311 0.01775005 0.05095575 0.121988
#> 391  0.07950311 0.01775005 0.05095575 0.121988
#> 392  0.07950311 0.01775005 0.05095575 0.121988
#> 393  0.07950311 0.01775005 0.05095575 0.121988
#> 394  0.07950311 0.01775005 0.05095575 0.121988
#> 395  0.07950311 0.01775005 0.05095575 0.121988
#> 396  0.07950311 0.01775005 0.05095575 0.121988
#> 397  0.07950311 0.01775005 0.05095575 0.121988
#> 398  0.07950311 0.01775005 0.05095575 0.121988
#> 399  0.07950311 0.01775005 0.05095575 0.121988
#> 400  0.07950311 0.01775005 0.05095575 0.121988
#> 401  0.07950311 0.01775005 0.05095575 0.121988
#> 402  0.07950311 0.01775005 0.05095575 0.121988
#> 403  0.07950311 0.01775005 0.05095575 0.121988
#> 404  0.07950311 0.01775005 0.05095575 0.121988
#> 405  0.07950311 0.01775005 0.05095575 0.121988
#> 406  0.07950311 0.01775005 0.05095575 0.121988
#> 407  0.07950311 0.01775005 0.05095575 0.121988
#> 408  0.07950311 0.01775005 0.05095575 0.121988
#> 409  0.07950311 0.01775005 0.05095575 0.121988
#> 410  0.07950311 0.01775005 0.05095575 0.121988
#> 411  0.07950311 0.01775005 0.05095575 0.121988
#> 412  0.07950311 0.01775005 0.05095575 0.121988
#> 413  0.07950311 0.01775005 0.05095575 0.121988
#> 414  0.07950311 0.01775005 0.05095575 0.121988
#> 415  0.07950311 0.01775005 0.05095575 0.121988
#> 416  0.07950311 0.01775005 0.05095575 0.121988
#> 417  0.07950311 0.01775005 0.05095575 0.121988
#> 418  0.07950311 0.01775005 0.05095575 0.121988
#> 419  0.07950311 0.01775005 0.05095575 0.121988
#> 420  0.07950311 0.01775005 0.05095575 0.121988
#> 421  0.07950311 0.01775005 0.05095575 0.121988
#> 422  0.07950311 0.01775005 0.05095575 0.121988
#> 423  0.07950311 0.01775005 0.05095575 0.121988
#> 424  0.07950311 0.01775005 0.05095575 0.121988
#> 425  0.07950311 0.01775005 0.05095575 0.121988
#> 426  0.07950311 0.01775005 0.05095575 0.121988
#> 427  0.07950311 0.01775005 0.05095575 0.121988
#> 428  0.07950311 0.01775005 0.05095575 0.121988
#> 429  0.07950311 0.01775005 0.05095575 0.121988
#> 430  0.07950311 0.01775005 0.05095575 0.121988
#> 431  0.07950311 0.01775005 0.05095575 0.121988
#> 432  0.07950311 0.01775005 0.05095575 0.121988
#> 433  0.07950311 0.01775005 0.05095575 0.121988
#> 434  0.07950311 0.01775005 0.05095575 0.121988
#> 435  0.07950311 0.01775005 0.05095575 0.121988
#> 436  0.07950311 0.01775005 0.05095575 0.121988
#> 437  0.07950311 0.01775005 0.05095575 0.121988
#> 438  0.07950311 0.01775005 0.05095575 0.121988
#> 439  0.07950311 0.01775005 0.05095575 0.121988
#> 440  0.07950311 0.01775005 0.05095575 0.121988
#> 441  0.07950311 0.01775005 0.05095575 0.121988
#> 442  0.07950311 0.01775005 0.05095575 0.121988
#> 443  0.07950311 0.01775005 0.05095575 0.121988
#> 444  0.07950311 0.01775005 0.05095575 0.121988
#> 445  0.07950311 0.01775005 0.05095575 0.121988
#> 446  0.07950311 0.01775005 0.05095575 0.121988
#> 447  0.07950311 0.01775005 0.05095575 0.121988
#> 448  0.07950311 0.01775005 0.05095575 0.121988
#> 449  0.07950311 0.01775005 0.05095575 0.121988
#> 450  0.07950311 0.01775005 0.05095575 0.121988
#> 451  0.07950311 0.01775005 0.05095575 0.121988
#> 452  0.07950311 0.01775005 0.05095575 0.121988
#> 453  0.07950311 0.01775005 0.05095575 0.121988
#> 454  0.07950311 0.01775005 0.05095575 0.121988
#> 455  0.07950311 0.01775005 0.05095575 0.121988
#> 456  0.07950311 0.01775005 0.05095575 0.121988
#> 457  0.07950311 0.01775005 0.05095575 0.121988
#> 458  0.07950311 0.01775005 0.05095575 0.121988
#> 459  0.07950311 0.01775005 0.05095575 0.121988
#> 460  0.07950311 0.01775005 0.05095575 0.121988
#> 461  0.07950311 0.01775005 0.05095575 0.121988
#> 462  0.07950311 0.01775005 0.05095575 0.121988
#> 463  0.07950311 0.01775005 0.05095575 0.121988
#> 464  0.07950311 0.01775005 0.05095575 0.121988
#> 465  0.07950311 0.01775005 0.05095575 0.121988
#> 466  0.07950311 0.01775005 0.05095575 0.121988
#> 467  0.07950311 0.01775005 0.05095575 0.121988
#> 468  0.07950311 0.01775005 0.05095575 0.121988
#> 469  0.07950311 0.01775005 0.05095575 0.121988
#> 470  0.07950311 0.01775005 0.05095575 0.121988
#> 471  0.07950311 0.01775005 0.05095575 0.121988
#> 472  0.07950311 0.01775005 0.05095575 0.121988
#> 473  0.07950311 0.01775005 0.05095575 0.121988
#> 474  0.07950311 0.01775005 0.05095575 0.121988
#> 475  0.07950311 0.01775005 0.05095575 0.121988
#> 476  0.07950311 0.01775005 0.05095575 0.121988
#> 477  0.07950311 0.01775005 0.05095575 0.121988
#> 478  0.07950311 0.01775005 0.05095575 0.121988
#> 479  0.07950311 0.01775005 0.05095575 0.121988
#> 480  0.07950311 0.01775005 0.05095575 0.121988
#> 481  0.07950311 0.01775005 0.05095575 0.121988
#> 482  0.07950311 0.01775005 0.05095575 0.121988
#> 483  0.07950311 0.01775005 0.05095575 0.121988
#> 484  0.07950311 0.01775005 0.05095575 0.121988
#> 485  0.07950311 0.01775005 0.05095575 0.121988
#> 486  0.07950311 0.01775005 0.05095575 0.121988
#> 487  0.07950311 0.01775005 0.05095575 0.121988
#> 488  0.07950311 0.01775005 0.05095575 0.121988
#> 489  0.07950311 0.01775005 0.05095575 0.121988
#> 490  0.07950311 0.01775005 0.05095575 0.121988
#> 491  0.07950311 0.01775005 0.05095575 0.121988
#> 492  0.07950311 0.01775005 0.05095575 0.121988
#> 493  0.07950311 0.01775005 0.05095575 0.121988
#> 494  0.07950311 0.01775005 0.05095575 0.121988
#> 495  0.07950311 0.01775005 0.05095575 0.121988
#> 496  0.07950311 0.01775005 0.05095575 0.121988
#> 497  0.07950311 0.01775005 0.05095575 0.121988
#> 498  0.07950311 0.01775005 0.05095575 0.121988
#> 499  0.07950311 0.01775005 0.05095575 0.121988
#> 500  0.07950311 0.01775005 0.05095575 0.121988
#> 501  0.07950311 0.01775005 0.05095575 0.121988
#> 502  0.07950311 0.01775005 0.05095575 0.121988
#> 503  0.07950311 0.01775005 0.05095575 0.121988
#> 504  0.07950311 0.01775005 0.05095575 0.121988
#> 505  0.07950311 0.01775005 0.05095575 0.121988
#> 506  0.07950311 0.01775005 0.05095575 0.121988
#> 507  0.07950311 0.01775005 0.05095575 0.121988
#> 508  0.07950311 0.01775005 0.05095575 0.121988
#> 509  0.07950311 0.01775005 0.05095575 0.121988
#> 510  0.07950311 0.01775005 0.05095575 0.121988
#> 511  0.07950311 0.01775005 0.05095575 0.121988
#> 512  0.07950311 0.01775005 0.05095575 0.121988
#> 513  0.07950311 0.01775005 0.05095575 0.121988
#> 514  0.07950311 0.01775005 0.05095575 0.121988
#> 515  0.07950311 0.01775005 0.05095575 0.121988
#> 516  0.07950311 0.01775005 0.05095575 0.121988
#> 517  0.07950311 0.01775005 0.05095575 0.121988
#> 518  0.07950311 0.01775005 0.05095575 0.121988
#> 519  0.07950311 0.01775005 0.05095575 0.121988
#> 520  0.07950311 0.01775005 0.05095575 0.121988
#> 521  0.07950311 0.01775005 0.05095575 0.121988
#> 522  0.07950311 0.01775005 0.05095575 0.121988
#> 523  0.07950311 0.01775005 0.05095575 0.121988
#> 524  0.07950311 0.01775005 0.05095575 0.121988
#> 525  0.07950311 0.01775005 0.05095575 0.121988
#> 526  0.07950311 0.01775005 0.05095575 0.121988
#> 527  0.07950311 0.01775005 0.05095575 0.121988
#> 528  0.07950311 0.01775005 0.05095575 0.121988
#> 529  0.07950311 0.01775005 0.05095575 0.121988
#> 530  0.07950311 0.01775005 0.05095575 0.121988
#> 531  0.07950311 0.01775005 0.05095575 0.121988
#> 532  0.07950311 0.01775005 0.05095575 0.121988
#> 533  0.07950311 0.01775005 0.05095575 0.121988
#> 534  0.07950311 0.01775005 0.05095575 0.121988
#> 535  0.07950311 0.01775005 0.05095575 0.121988
#> 536  0.07950311 0.01775005 0.05095575 0.121988
#> 537  0.07950311 0.01775005 0.05095575 0.121988
#> 538  0.07950311 0.01775005 0.05095575 0.121988
#> 539  0.07950311 0.01775005 0.05095575 0.121988
#> 540  0.07950311 0.01775005 0.05095575 0.121988
#> 541  0.07950311 0.01775005 0.05095575 0.121988
#> 542  0.07950311 0.01775005 0.05095575 0.121988
#> 543  0.07950311 0.01775005 0.05095575 0.121988
#> 544  0.07950311 0.01775005 0.05095575 0.121988
#> 545  0.07950311 0.01775005 0.05095575 0.121988
#> 546  0.07950311 0.01775005 0.05095575 0.121988
#> 547  0.07950311 0.01775005 0.05095575 0.121988
#> 548  0.07950311 0.01775005 0.05095575 0.121988
#> 549  0.07950311 0.01775005 0.05095575 0.121988
#> 550  0.07950311 0.01775005 0.05095575 0.121988
#> 551  0.07950311 0.01775005 0.05095575 0.121988
#> 552  0.07950311 0.01775005 0.05095575 0.121988
#> 553  0.07950311 0.01775005 0.05095575 0.121988
#> 554  0.07950311 0.01775005 0.05095575 0.121988
#> 555  0.07950311 0.01775005 0.05095575 0.121988
#> 556  0.07950311 0.01775005 0.05095575 0.121988
#> 557  0.07950311 0.01775005 0.05095575 0.121988
#> 558  0.07950311 0.01775005 0.05095575 0.121988
#> 559  0.07950311 0.01775005 0.05095575 0.121988
#> 560  0.07950311 0.01775005 0.05095575 0.121988
#> 561  0.07950311 0.01775005 0.05095575 0.121988
#> 562  0.07950311 0.01775005 0.05095575 0.121988
#> 563  0.07950311 0.01775005 0.05095575 0.121988
#> 564  0.07950311 0.01775005 0.05095575 0.121988
#> 565  0.07950311 0.01775005 0.05095575 0.121988
#> 566  0.07950311 0.01775005 0.05095575 0.121988
#> 567  0.07950311 0.01775005 0.05095575 0.121988
#> 568  0.07950311 0.01775005 0.05095575 0.121988
#> 569  0.07950311 0.01775005 0.05095575 0.121988
#> 570  0.07950311 0.01775005 0.05095575 0.121988
#> 571  0.07950311 0.01775005 0.05095575 0.121988
#> 572  0.07950311 0.01775005 0.05095575 0.121988
#> 573  0.07950311 0.01775005 0.05095575 0.121988
#> 574  0.07950311 0.01775005 0.05095575 0.121988
#> 575  0.07950311 0.01775005 0.05095575 0.121988
#> 576  0.07950311 0.01775005 0.05095575 0.121988
#> 577  0.07950311 0.01775005 0.05095575 0.121988
#> 578  0.07950311 0.01775005 0.05095575 0.121988
#> 579  0.07950311 0.01775005 0.05095575 0.121988
#> 580  0.07950311 0.01775005 0.05095575 0.121988
#> 581  0.07950311 0.01775005 0.05095575 0.121988
#> 582  0.07950311 0.01775005 0.05095575 0.121988
#> 583  0.07950311 0.01775005 0.05095575 0.121988
#> 584  0.07950311 0.01775005 0.05095575 0.121988
#> 585  0.07950311 0.01775005 0.05095575 0.121988
#> 586  0.07950311 0.01775005 0.05095575 0.121988
#> 587  0.07950311 0.01775005 0.05095575 0.121988
#> 588  0.07950311 0.01775005 0.05095575 0.121988
#> 589  0.07950311 0.01775005 0.05095575 0.121988
#> 590  0.07950311 0.01775005 0.05095575 0.121988
#> 591  0.07950311 0.01775005 0.05095575 0.121988
#> 592  0.07950311 0.01775005 0.05095575 0.121988
#> 593  0.07950311 0.01775005 0.05095575 0.121988
#> 594  0.07950311 0.01775005 0.05095575 0.121988
#> 595  0.07950311 0.01775005 0.05095575 0.121988
#> 596  0.07950311 0.01775005 0.05095575 0.121988
#> 597  0.07950311 0.01775005 0.05095575 0.121988
#> 598  0.07950311 0.01775005 0.05095575 0.121988
#> 599  0.07950311 0.01775005 0.05095575 0.121988
#> 600  0.07950311 0.01775005 0.05095575 0.121988
#> 601  0.07950311 0.01775005 0.05095575 0.121988
#> 602  0.07950311 0.01775005 0.05095575 0.121988
#> 603  0.07950311 0.01775005 0.05095575 0.121988
#> 604  0.07950311 0.01775005 0.05095575 0.121988
#> 605  0.07950311 0.01775005 0.05095575 0.121988
#> 606  0.07950311 0.01775005 0.05095575 0.121988
#> 607  0.07950311 0.01775005 0.05095575 0.121988
#> 608  0.07950311 0.01775005 0.05095575 0.121988
#> 609  0.07950311 0.01775005 0.05095575 0.121988
#> 610  0.07950311 0.01775005 0.05095575 0.121988
#> 611  0.07950311 0.01775005 0.05095575 0.121988
#> 612  0.07950311 0.01775005 0.05095575 0.121988
#> 613  0.07950311 0.01775005 0.05095575 0.121988
#> 614  0.07950311 0.01775005 0.05095575 0.121988
#> 615  0.07950311 0.01775005 0.05095575 0.121988
#> 616  0.07950311 0.01775005 0.05095575 0.121988
#> 617  0.07950311 0.01775005 0.05095575 0.121988
#> 618  0.07950311 0.01775005 0.05095575 0.121988
#> 619  0.07950311 0.01775005 0.05095575 0.121988
#> 620  0.07950311 0.01775005 0.05095575 0.121988
#> 621  0.07950311 0.01775005 0.05095575 0.121988
#> 622  0.07950311 0.01775005 0.05095575 0.121988
#> 623  0.07950311 0.01775005 0.05095575 0.121988
#> 624  0.07950311 0.01775005 0.05095575 0.121988
#> 625  0.07950311 0.01775005 0.05095575 0.121988
#> 626  0.07950311 0.01775005 0.05095575 0.121988
#> 627  0.07950311 0.01775005 0.05095575 0.121988
#> 628  0.07950311 0.01775005 0.05095575 0.121988
#> 629  0.07950311 0.01775005 0.05095575 0.121988
#> 630  0.07950311 0.01775005 0.05095575 0.121988
#> 631  0.07950311 0.01775005 0.05095575 0.121988
#> 632  0.07950311 0.01775005 0.05095575 0.121988
#> 633  0.07950311 0.01775005 0.05095575 0.121988
#> 634  0.07950311 0.01775005 0.05095575 0.121988
#> 635  0.07950311 0.01775005 0.05095575 0.121988
#> 636  0.07950311 0.01775005 0.05095575 0.121988
#> 637  0.07950311 0.01775005 0.05095575 0.121988
#> 638  0.07950311 0.01775005 0.05095575 0.121988
#> 639  0.07950311 0.01775005 0.05095575 0.121988
#> 640  0.07950311 0.01775005 0.05095575 0.121988
#> 641  0.07950311 0.01775005 0.05095575 0.121988
#> 642  0.07950311 0.01775005 0.05095575 0.121988
#> 643  0.07950311 0.01775005 0.05095575 0.121988
#> 644  0.07950311 0.01775005 0.05095575 0.121988
#> 645  0.07950311 0.01775005 0.05095575 0.121988
#> 646  0.07950311 0.01775005 0.05095575 0.121988
#> 647  0.07950311 0.01775005 0.05095575 0.121988
#> 648  0.07950311 0.01775005 0.05095575 0.121988
#> 649  0.07950311 0.01775005 0.05095575 0.121988
#> 650  0.07950311 0.01775005 0.05095575 0.121988
#> 651  0.07950311 0.01775005 0.05095575 0.121988
#> 652  0.07950311 0.01775005 0.05095575 0.121988
#> 653  0.07950311 0.01775005 0.05095575 0.121988
#> 654  0.07950311 0.01775005 0.05095575 0.121988
#> 655  0.07950311 0.01775005 0.05095575 0.121988
#> 656  0.07950311 0.01775005 0.05095575 0.121988
#> 657  0.07950311 0.01775005 0.05095575 0.121988
#> 658  0.07950311 0.01775005 0.05095575 0.121988
#> 659  0.07950311 0.01775005 0.05095575 0.121988
#> 660  0.07950311 0.01775005 0.05095575 0.121988
#> 661  0.07950311 0.01775005 0.05095575 0.121988
#> 662  0.07950311 0.01775005 0.05095575 0.121988
#> 663  0.07950311 0.01775005 0.05095575 0.121988
#> 664  0.07950311 0.01775005 0.05095575 0.121988
#> 665  0.07950311 0.01775005 0.05095575 0.121988
#> 666  0.07950311 0.01775005 0.05095575 0.121988
#> 667  0.07950311 0.01775005 0.05095575 0.121988
#> 668  0.07950311 0.01775005 0.05095575 0.121988
#> 669  0.07950311 0.01775005 0.05095575 0.121988
#> 670  0.07950311 0.01775005 0.05095575 0.121988
#> 671  0.07950311 0.01775005 0.05095575 0.121988
#> 672  0.07950311 0.01775005 0.05095575 0.121988
#> 673  0.07950311 0.01775005 0.05095575 0.121988
#> 674  0.07950311 0.01775005 0.05095575 0.121988
#> 675  0.07950311 0.01775005 0.05095575 0.121988
#> 676  0.07950311 0.01775005 0.05095575 0.121988
#> 677  0.07950311 0.01775005 0.05095575 0.121988
#> 678  0.07950311 0.01775005 0.05095575 0.121988
#> 679  0.07950311 0.01775005 0.05095575 0.121988
#> 680  0.07950311 0.01775005 0.05095575 0.121988
#> 681  0.07950311 0.01775005 0.05095575 0.121988
#> 682  0.07950311 0.01775005 0.05095575 0.121988
#> 683  0.07950311 0.01775005 0.05095575 0.121988
#> 684  0.07950311 0.01775005 0.05095575 0.121988
#> 685  0.07950311 0.01775005 0.05095575 0.121988
#> 686  0.07950311 0.01775005 0.05095575 0.121988
#> 687  0.07950311 0.01775005 0.05095575 0.121988
#> 688  0.07950311 0.01775005 0.05095575 0.121988
#> 689  0.07950311 0.01775005 0.05095575 0.121988
#> 690  0.07950311 0.01775005 0.05095575 0.121988
#> 691  0.07950311 0.01775005 0.05095575 0.121988
#> 692  0.07950311 0.01775005 0.05095575 0.121988
#> 693  0.07950311 0.01775005 0.05095575 0.121988
#> 694  0.07950311 0.01775005 0.05095575 0.121988
#> 695  0.07950311 0.01775005 0.05095575 0.121988
#> 696  0.07950311 0.01775005 0.05095575 0.121988
#> 697  0.07950311 0.01775005 0.05095575 0.121988
#> 698  0.07950311 0.01775005 0.05095575 0.121988
#> 699  0.07950311 0.01775005 0.05095575 0.121988
#> 700  0.07950311 0.01775005 0.05095575 0.121988
#> 701  0.07950311 0.01775005 0.05095575 0.121988
#> 702  0.07950311 0.01775005 0.05095575 0.121988
#> 703  0.07950311 0.01775005 0.05095575 0.121988
#> 704  0.07950311 0.01775005 0.05095575 0.121988
#> 705  0.07950311 0.01775005 0.05095575 0.121988
#> 706  0.07950311 0.01775005 0.05095575 0.121988
#> 707  0.07950311 0.01775005 0.05095575 0.121988
#> 708  0.07950311 0.01775005 0.05095575 0.121988
#> 709  0.07950311 0.01775005 0.05095575 0.121988
#> 710  0.07950311 0.01775005 0.05095575 0.121988
#> 711  0.07950311 0.01775005 0.05095575 0.121988
#> 712  0.07950311 0.01775005 0.05095575 0.121988
#> 713  0.07950311 0.01775005 0.05095575 0.121988
#> 714  0.07950311 0.01775005 0.05095575 0.121988
#> 715  0.07950311 0.01775005 0.05095575 0.121988
#> 716  0.07950311 0.01775005 0.05095575 0.121988
#> 717  0.07950311 0.01775005 0.05095575 0.121988
#> 718  0.07950311 0.01775005 0.05095575 0.121988
#> 719  0.07950311 0.01775005 0.05095575 0.121988
#> 720  0.07950311 0.01775005 0.05095575 0.121988
#> 721  0.07950311 0.01775005 0.05095575 0.121988
#> 722  0.07950311 0.01775005 0.05095575 0.121988
#> 723  0.07950311 0.01775005 0.05095575 0.121988
#> 724  0.07950311 0.01775005 0.05095575 0.121988
#> 725  0.07950311 0.01775005 0.05095575 0.121988
#> 726  0.07950311 0.01775005 0.05095575 0.121988
#> 727  0.07950311 0.01775005 0.05095575 0.121988
#> 728  0.07950311 0.01775005 0.05095575 0.121988
#> 729  0.07950311 0.01775005 0.05095575 0.121988
#> 730  0.07950311 0.01775005 0.05095575 0.121988
#> 731  0.07950311 0.01775005 0.05095575 0.121988
#> 732  0.07950311 0.01775005 0.05095575 0.121988
#> 733  0.07950311 0.01775005 0.05095575 0.121988
#> 734  0.07950311 0.01775005 0.05095575 0.121988
#> 735  0.07950311 0.01775005 0.05095575 0.121988
#> 736  0.07950311 0.01775005 0.05095575 0.121988
#> 737  0.07950311 0.01775005 0.05095575 0.121988
#> 738  0.07950311 0.01775005 0.05095575 0.121988
#> 739  0.07950311 0.01775005 0.05095575 0.121988
#> 740  0.07950311 0.01775005 0.05095575 0.121988
#> 741  0.07950311 0.01775005 0.05095575 0.121988
#> 742  0.07950311 0.01775005 0.05095575 0.121988
#> 743  0.07950311 0.01775005 0.05095575 0.121988
#> 744  0.07950311 0.01775005 0.05095575 0.121988
#> 745  0.07950311 0.01775005 0.05095575 0.121988
#> 746  0.07950311 0.01775005 0.05095575 0.121988
#> 747  0.07950311 0.01775005 0.05095575 0.121988
#> 748  0.07950311 0.01775005 0.05095575 0.121988
#> 749  0.07950311 0.01775005 0.05095575 0.121988
#> 750  0.07950311 0.01775005 0.05095575 0.121988
#> 751  0.07950311 0.01775005 0.05095575 0.121988
#> 752  0.07950311 0.01775005 0.05095575 0.121988
#> 753  0.07950311 0.01775005 0.05095575 0.121988
#> 754  0.07950311 0.01775005 0.05095575 0.121988
#> 755  0.07950311 0.01775005 0.05095575 0.121988
#> 756  0.07950311 0.01775005 0.05095575 0.121988
#> 757  0.07950311 0.01775005 0.05095575 0.121988
#> 758  0.07950311 0.01775005 0.05095575 0.121988
#> 759  0.07950311 0.01775005 0.05095575 0.121988
#> 760  0.07950311 0.01775005 0.05095575 0.121988
#> 761  0.07950311 0.01775005 0.05095575 0.121988
#> 762  0.07950311 0.01775005 0.05095575 0.121988
#> 763  0.07950311 0.01775005 0.05095575 0.121988
#> 764  0.07950311 0.01775005 0.05095575 0.121988
#> 765  0.07950311 0.01775005 0.05095575 0.121988
#> 766  0.07950311 0.01775005 0.05095575 0.121988
#> 767  0.07950311 0.01775005 0.05095575 0.121988
#> 768  0.07950311 0.01775005 0.05095575 0.121988
#> 769  0.07950311 0.01775005 0.05095575 0.121988
#> 770  0.07950311 0.01775005 0.05095575 0.121988
#> 771  0.07950311 0.01775005 0.05095575 0.121988
#> 772  0.07950311 0.01775005 0.05095575 0.121988
#> 773  0.07950311 0.01775005 0.05095575 0.121988
#> 774  0.07950311 0.01775005 0.05095575 0.121988
#> 775  0.07950311 0.01775005 0.05095575 0.121988
#> 776  0.07950311 0.01775005 0.05095575 0.121988
#> 777  0.07950311 0.01775005 0.05095575 0.121988
#> 778  0.07950311 0.01775005 0.05095575 0.121988
#> 779  0.07950311 0.01775005 0.05095575 0.121988
#> 780  0.07950311 0.01775005 0.05095575 0.121988
#> 781  0.07950311 0.01775005 0.05095575 0.121988
#> 782  0.07950311 0.01775005 0.05095575 0.121988
#> 783  0.07950311 0.01775005 0.05095575 0.121988
#> 784  0.07950311 0.01775005 0.05095575 0.121988
#> 785  0.07950311 0.01775005 0.05095575 0.121988
#> 786  0.07950311 0.01775005 0.05095575 0.121988
#> 787  0.07950311 0.01775005 0.05095575 0.121988
#> 788  0.07950311 0.01775005 0.05095575 0.121988
#> 789  0.07950311 0.01775005 0.05095575 0.121988
#> 790  0.07950311 0.01775005 0.05095575 0.121988
#> 791  0.07950311 0.01775005 0.05095575 0.121988
#> 792  0.07950311 0.01775005 0.05095575 0.121988
#> 793  0.07950311 0.01775005 0.05095575 0.121988
#> 794  0.07950311 0.01775005 0.05095575 0.121988
#> 795  0.07950311 0.01775005 0.05095575 0.121988
#> 796  0.07950311 0.01775005 0.05095575 0.121988
#> 797  0.07950311 0.01775005 0.05095575 0.121988
#> 798  0.07950311 0.01775005 0.05095575 0.121988
#> 799  0.07950311 0.01775005 0.05095575 0.121988
#> 800  0.07950311 0.01775005 0.05095575 0.121988
#> 801  0.07950311 0.01775005 0.05095575 0.121988
#> 802  0.07950311 0.01775005 0.05095575 0.121988
#> 803  0.07950311 0.01775005 0.05095575 0.121988
#> 804  0.07950311 0.01775005 0.05095575 0.121988
#> 805  0.07950311 0.01775005 0.05095575 0.121988
#> 806  0.07950311 0.01775005 0.05095575 0.121988
#> 807  0.07950311 0.01775005 0.05095575 0.121988
#> 808  0.07950311 0.01775005 0.05095575 0.121988
#> 809  0.07950311 0.01775005 0.05095575 0.121988
#> 810  0.07950311 0.01775005 0.05095575 0.121988
#> 811  0.07950311 0.01775005 0.05095575 0.121988
#> 812  0.07950311 0.01775005 0.05095575 0.121988
#> 813  0.07950311 0.01775005 0.05095575 0.121988
#> 814  0.07950311 0.01775005 0.05095575 0.121988
#> 815  0.07950311 0.01775005 0.05095575 0.121988
#> 816  0.07950311 0.01775005 0.05095575 0.121988
#> 817  0.07950311 0.01775005 0.05095575 0.121988
#> 818  0.07950311 0.01775005 0.05095575 0.121988
#> 819  0.07950311 0.01775005 0.05095575 0.121988
#> 820  0.07950311 0.01775005 0.05095575 0.121988
#> 821  0.07950311 0.01775005 0.05095575 0.121988
#> 822  0.07950311 0.01775005 0.05095575 0.121988
#> 823  0.07950311 0.01775005 0.05095575 0.121988
#> 824  0.07950311 0.01775005 0.05095575 0.121988
#> 825  0.07950311 0.01775005 0.05095575 0.121988
#> 826  0.07950311 0.01775005 0.05095575 0.121988
#> 827  0.07950311 0.01775005 0.05095575 0.121988
#> 828  0.07950311 0.01775005 0.05095575 0.121988
#> 829  0.07950311 0.01775005 0.05095575 0.121988
#> 830  0.07950311 0.01775005 0.05095575 0.121988
#> 831  0.07950311 0.01775005 0.05095575 0.121988
#> 832  0.07950311 0.01775005 0.05095575 0.121988
#> 833  0.07950311 0.01775005 0.05095575 0.121988
#> 834  0.07950311 0.01775005 0.05095575 0.121988
#> 835  0.07950311 0.01775005 0.05095575 0.121988
#> 836  0.07950311 0.01775005 0.05095575 0.121988
#> 837  0.07950311 0.01775005 0.05095575 0.121988
#> 838  0.07950311 0.01775005 0.05095575 0.121988
#> 839  0.07950311 0.01775005 0.05095575 0.121988
#> 840  0.07950311 0.01775005 0.05095575 0.121988
#> 841  0.07950311 0.01775005 0.05095575 0.121988
#> 842  0.07950311 0.01775005 0.05095575 0.121988
#> 843  0.07950311 0.01775005 0.05095575 0.121988
#> 844  0.07950311 0.01775005 0.05095575 0.121988
#> 845  0.07950311 0.01775005 0.05095575 0.121988
#> 846  0.07950311 0.01775005 0.05095575 0.121988
#> 847  0.07950311 0.01775005 0.05095575 0.121988
#> 848  0.07950311 0.01775005 0.05095575 0.121988
#> 849  0.07950311 0.01775005 0.05095575 0.121988
#> 850  0.07950311 0.01775005 0.05095575 0.121988
#> 851  0.07950311 0.01775005 0.05095575 0.121988
#> 852  0.07950311 0.01775005 0.05095575 0.121988
#> 853  0.07950311 0.01775005 0.05095575 0.121988
#> 854  0.07950311 0.01775005 0.05095575 0.121988
#> 855  0.07950311 0.01775005 0.05095575 0.121988
#> 856  0.07950311 0.01775005 0.05095575 0.121988
#> 857  0.07950311 0.01775005 0.05095575 0.121988
#> 858  0.07950311 0.01775005 0.05095575 0.121988
#> 859  0.07950311 0.01775005 0.05095575 0.121988
#> 860  0.07950311 0.01775005 0.05095575 0.121988
#> 861  0.07950311 0.01775005 0.05095575 0.121988
#> 862  0.07950311 0.01775005 0.05095575 0.121988
#> 863  0.07950311 0.01775005 0.05095575 0.121988
#> 864  0.07950311 0.01775005 0.05095575 0.121988
#> 865  0.07950311 0.01775005 0.05095575 0.121988
#> 866  0.07950311 0.01775005 0.05095575 0.121988
#> 867  0.07950311 0.01775005 0.05095575 0.121988
#> 868  0.07950311 0.01775005 0.05095575 0.121988
#> 869  0.07950311 0.01775005 0.05095575 0.121988
#> 870  0.07950311 0.01775005 0.05095575 0.121988
#> 871  0.07950311 0.01775005 0.05095575 0.121988
#> 872  0.07950311 0.01775005 0.05095575 0.121988
#> 873  0.07950311 0.01775005 0.05095575 0.121988
#> 874  0.07950311 0.01775005 0.05095575 0.121988
#> 875  0.07950311 0.01775005 0.05095575 0.121988
#> 876  0.07950311 0.01775005 0.05095575 0.121988
#> 877  0.07950311 0.01775005 0.05095575 0.121988
#> 878  0.07950311 0.01775005 0.05095575 0.121988
#> 879  0.07950311 0.01775005 0.05095575 0.121988
#> 880  0.07950311 0.01775005 0.05095575 0.121988
#> 881  0.07950311 0.01775005 0.05095575 0.121988
#> 882  0.07950311 0.01775005 0.05095575 0.121988
#> 883  0.07950311 0.01775005 0.05095575 0.121988
#> 884  0.07950311 0.01775005 0.05095575 0.121988
#> 885  0.07950311 0.01775005 0.05095575 0.121988
#> 886  0.07950311 0.01775005 0.05095575 0.121988
#> 887  0.07950311 0.01775005 0.05095575 0.121988
#> 888  0.07950311 0.01775005 0.05095575 0.121988
#> 889  0.07950311 0.01775005 0.05095575 0.121988
#> 890  0.07950311 0.01775005 0.05095575 0.121988
#> 891  0.07950311 0.01775005 0.05095575 0.121988
#> 892  0.07950311 0.01775005 0.05095575 0.121988
#> 893  0.07950311 0.01775005 0.05095575 0.121988
#> 894  0.07950311 0.01775005 0.05095575 0.121988
#> 895  0.07950311 0.01775005 0.05095575 0.121988
#> 896  0.07950311 0.01775005 0.05095575 0.121988
#> 897  0.07950311 0.01775005 0.05095575 0.121988
#> 898  0.07950311 0.01775005 0.05095575 0.121988
#> 899  0.07950311 0.01775005 0.05095575 0.121988
#> 900  0.07950311 0.01775005 0.05095575 0.121988
#> 901  0.07950311 0.01775005 0.05095575 0.121988
#> 902  0.07950311 0.01775005 0.05095575 0.121988
#> 903  0.07950311 0.01775005 0.05095575 0.121988
#> 904  0.07950311 0.01775005 0.05095575 0.121988
#> 905  0.07950311 0.01775005 0.05095575 0.121988
#> 906  0.07950311 0.01775005 0.05095575 0.121988
#> 907  0.07950311 0.01775005 0.05095575 0.121988
#> 908  0.07950311 0.01775005 0.05095575 0.121988
#> 909  0.07950311 0.01775005 0.05095575 0.121988
#> 910  0.07950311 0.01775005 0.05095575 0.121988
#> 911  0.07950311 0.01775005 0.05095575 0.121988
#> 912  0.07950311 0.01775005 0.05095575 0.121988
#> 913  0.07950311 0.01775005 0.05095575 0.121988
#> 914  0.07950311 0.01775005 0.05095575 0.121988
#> 915  0.07950311 0.01775005 0.05095575 0.121988
#> 916  0.07950311 0.01775005 0.05095575 0.121988
#> 917  0.07950311 0.01775005 0.05095575 0.121988
#> 918  0.07950311 0.01775005 0.05095575 0.121988
#> 919  0.07950311 0.01775005 0.05095575 0.121988
#> 920  0.07950311 0.01775005 0.05095575 0.121988
#> 921  0.07950311 0.01775005 0.05095575 0.121988
#> 922  0.07950311 0.01775005 0.05095575 0.121988
#> 923  0.07950311 0.01775005 0.05095575 0.121988
#> 924  0.07950311 0.01775005 0.05095575 0.121988
#> 925  0.07950311 0.01775005 0.05095575 0.121988
#> 926  0.07950311 0.01775005 0.05095575 0.121988
#> 927  0.07950311 0.01775005 0.05095575 0.121988
#> 928  0.07950311 0.01775005 0.05095575 0.121988
#> 929  0.07950311 0.01775005 0.05095575 0.121988
#> 930  0.07950311 0.01775005 0.05095575 0.121988
#> 931  0.07950311 0.01775005 0.05095575 0.121988
#> 932  0.07950311 0.01775005 0.05095575 0.121988
#> 933  0.07950311 0.01775005 0.05095575 0.121988
#> 934  0.07950311 0.01775005 0.05095575 0.121988
#> 935  0.07950311 0.01775005 0.05095575 0.121988
#> 936  0.07950311 0.01775005 0.05095575 0.121988
#> 937  0.07950311 0.01775005 0.05095575 0.121988
#> 938  0.07950311 0.01775005 0.05095575 0.121988
#> 939  0.07950311 0.01775005 0.05095575 0.121988
#> 940  0.07950311 0.01775005 0.05095575 0.121988
#> 941  0.07950311 0.01775005 0.05095575 0.121988
#> 942  0.07950311 0.01775005 0.05095575 0.121988
#> 943  0.07950311 0.01775005 0.05095575 0.121988
#> 944  0.07950311 0.01775005 0.05095575 0.121988
#> 945  0.07950311 0.01775005 0.05095575 0.121988
#> 946  0.07950311 0.01775005 0.05095575 0.121988
#> 947  0.07950311 0.01775005 0.05095575 0.121988
#> 948  0.07950311 0.01775005 0.05095575 0.121988
#> 949  0.07950311 0.01775005 0.05095575 0.121988
#> 950  0.07950311 0.01775005 0.05095575 0.121988
#> 951  0.07950311 0.01775005 0.05095575 0.121988
#> 952  0.07950311 0.01775005 0.05095575 0.121988
#> 953  0.07950311 0.01775005 0.05095575 0.121988
#> 954  0.07950311 0.01775005 0.05095575 0.121988
#> 955  0.07950311 0.01775005 0.05095575 0.121988
#> 956  0.07950311 0.01775005 0.05095575 0.121988
#> 957  0.07950311 0.01775005 0.05095575 0.121988
#> 958  0.07950311 0.01775005 0.05095575 0.121988
#> 959  0.07950311 0.01775005 0.05095575 0.121988
#> 960  0.07950311 0.01775005 0.05095575 0.121988
#> 961  0.07950311 0.01775005 0.05095575 0.121988
#> 962  0.07950311 0.01775005 0.05095575 0.121988
#> 963  0.07950311 0.01775005 0.05095575 0.121988
#> 964  0.07950311 0.01775005 0.05095575 0.121988
#> 965  0.07950311 0.01775005 0.05095575 0.121988
#> 966  0.07950311 0.01775005 0.05095575 0.121988
#> 967  0.07950311 0.01775005 0.05095575 0.121988
#> 968  0.07950311 0.01775005 0.05095575 0.121988
#> 969  0.07950311 0.01775005 0.05095575 0.121988
#> 970  0.07950311 0.01775005 0.05095575 0.121988
#> 971  0.07950311 0.01775005 0.05095575 0.121988
#> 972  0.07950311 0.01775005 0.05095575 0.121988
#> 973  0.07950311 0.01775005 0.05095575 0.121988
#> 974  0.07950311 0.01775005 0.05095575 0.121988
#> 975  0.07950311 0.01775005 0.05095575 0.121988
#> 976  0.07950311 0.01775005 0.05095575 0.121988
#> 977  0.07950311 0.01775005 0.05095575 0.121988
#> 978  0.07950311 0.01775005 0.05095575 0.121988
#> 979  0.07950311 0.01775005 0.05095575 0.121988
#> 980  0.07950311 0.01775005 0.05095575 0.121988
#> 981  0.07950311 0.01775005 0.05095575 0.121988
#> 982  0.07950311 0.01775005 0.05095575 0.121988
#> 983  0.07950311 0.01775005 0.05095575 0.121988
#> 984  0.07950311 0.01775005 0.05095575 0.121988
#> 985  0.07950311 0.01775005 0.05095575 0.121988
#> 986  0.07950311 0.01775005 0.05095575 0.121988
#> 987  0.07950311 0.01775005 0.05095575 0.121988
#> 988  0.07950311 0.01775005 0.05095575 0.121988
#> 989  0.07950311 0.01775005 0.05095575 0.121988
#> 990  0.07950311 0.01775005 0.05095575 0.121988
#> 991  0.07950311 0.01775005 0.05095575 0.121988
#> 992  0.07950311 0.01775005 0.05095575 0.121988
#> 993  0.07950311 0.01775005 0.05095575 0.121988
#> 994  0.07950311 0.01775005 0.05095575 0.121988
#> 995  0.07950311 0.01775005 0.05095575 0.121988
#> 996  0.07950311 0.01775005 0.05095575 0.121988
#> 997  0.07950311 0.01775005 0.05095575 0.121988
#> 998  0.07950311 0.01775005 0.05095575 0.121988
#> 999  0.07950311 0.01775005 0.05095575 0.121988
#> 1000 0.07950311 0.01775005 0.05095575 0.121988

coef(m1)

#>     psi(Int) psi(habitat)       p(Int)      p(METH)      fp(Int) 
#>   -1.3730028    2.6988048   -0.2436260    0.9657119   -2.4491175 

confint(m1, type = 'det')

#>              0.025     0.975
#> p(Int)  -0.5010582 0.0138062
#> p(METH)  0.6000841 1.3313396