Fit the occupancy model of Royle and Nichols (2003)

Fit the occupancy model of Royle and Nichols (2003), which relates probability of detection of the species to the number of individuals available for detection at each site. Probability of occupancy is a derived parameter: the probability that at least one individual is available for detection at the site.

occuRN(formula, data, K=25, starts, method="BFGS", se=TRUE, 
              engine=c("C","R"), threads=1, ...)

Arguments

formula
double right-hand side formula describing covariates of detection and abundance, in that order.

data
Object of class unmarkedFrameOccu supplying data to the model.

K
the upper summation index used to numerically integrate out the latent abundance. This should be set high enough so that it does not affect the parameter estimates. Computation time will increase with K.

starts
initial values for the optimization.

method
Optimization method used by optim.

se
logical specifying whether or not to compute standard errors.

engine
Either "C" to use fast C++ code or "R" to use native R code during the optimization.

threads
Set the number of threads to use for optimization in C++, if OpenMP is available on your system. Increasing the number of threads may speed up optimization in some cases by running the likelihood calculation in parallel. If threads=1 (the default), OpenMP is disabled.

...
Additional arguments to optim, such as lower and upper bounds

Details

This function fits the latent abundance mixture model described in Royle and Nichols (2003).

The number of animals available for detection at site \(i\) is modelled as Poisson:

$$N_i \sim Poisson(\lambda_i)$$

We assume that all individuals at site \(i\) during sample \(j\) have identical detection probabilities, \(r_{ij}\), and that detections are independent. The species will be recorded if at least one individual is detected. Thus, the detection probability for the species is linked to the detection probability for an individual by

$$p_{ij} = 1 - (1 - r_{ij}) ^ {N_i}$$

Note that if \(N_i = 0\), then \(p_{ij} = 0\), and increasing values of \(N_i\) lead to higher values of \(p_{ij}\) The equation for the detection history is then:

$$y_{ij} \sim Bernoulli(p_{ij})$$

Covariates of \(\lambda_i\) are modelled with the log link and covariates of \(r_{ij}\) are modelled with the logit link.

Value

unmarkedFit object describing the model fit.

Author

Ian Fiske

References

Royle, J. A. and Nichols, J. D. (2003) Estimating Abundance from Repeated Presence-Absence Data or Point Counts. Ecology, 84(3) pp. 777--790.

Examples

if (FALSE) {

data(birds)
woodthrushUMF <- unmarkedFrameOccu(woodthrush.bin)
# survey occasion-specific detection probabilities
(fm.wood.rn <- occuRN(~ obsNum ~ 1, woodthrushUMF))

# Empirical Bayes estimates of abundance at each site
re <- ranef(fm.wood.rn)
plot(re)

}