Fit the combined distance and removal model of Amundson et al. (2014).

Fit the model of Amundson et al. (2014) to point count datasets containing both distance and time of observation data. The Amundson et al. (2014) model is extended to account for temporary emigration by estimating an additional availability probability if multiple counts at a site are available. Abundance can be modeled as a Poisson, negative binomial, or Zero-inflated Poisson. Multiple distance sampling key functions are also available.

gdistremoval(lambdaformula=~1, phiformula=~1, removalformula=~1,
  distanceformula=~1, data, keyfun=c("halfnorm", "exp", "hazard", "uniform"),
  output=c("abund", "density"), unitsOut=c("ha", "kmsq"), mixture=c('P', 'NB', 'ZIP'), 
  K, starts, method = "BFGS", se = TRUE, engine=c("C","TMB"), threads=1, ...)

Arguments

lambdaformula

A right-hand side formula describing the abundance covariates

phiformula

A right-hand side formula describing the availability covariates

removalformula

A right-hand side formula describing removal probability covariates

distanceformula

A right-hand side formula describing the detection function covariates

data

An object of class unmarkedFrameGDR

keyfun

One of the following detection functions: "halfnorm", "hazard", "exp", or "uniform"

output

Model either "abund" or "density"

unitsOut

Units of density. Either "ha" or "kmsq" for hectares and square kilometers, respectively

mixture

Either "P", "NB", or "ZIP" for the Poisson, negative binomial, and Zero-inflated Poisson models of abundance

K

An integer value specifying the upper bound used in the integration

starts

A numeric vector of starting values for the model parameters

method

Optimization method used by optim

se

logical specifying whether or not to compute standard errors

engine

Either "C" to use C++ code or "TMB" to use TMB for optimization

threads

Set the number of threads to use for optimization in C++, if OpenMP is available on your system. Increasing the number of threads may speed up optimization in some cases by running the likelihood calculation in parallel. If threads=1 (the default), OpenMP is disabled

...

Additional arguments to optim, such as lower and upper bounds

Value

An object of class unmarkedFitGDR

References

Amundson, C.L., Royle, J.A. and Handel, C.M., 2014. A hierarchical model combining distance sampling and time removal to estimate detection probability during avian point counts. The Auk 131: 476-494.

Author

Ken Kellner contact@kenkellner.com

See also